+++ lib/marquee_direction.dart
@@ -1,0 +1,4 @@
+enum MarqueeDirection {
+  rtl,
+  ltr,
+}